+++ ubuntu-apache2-modphp-dev-7.0/startup-php-env-vars.sh
@@ -1,0 +1,31 @@
+#!/bin/bash
+
+# defaults that may be set in Dockerfile
+#
+# why do we have two sets?
+# - PHP_DEFAULT_* are for setting in Dockerfiles
+# - when I've attempted to override them in docker-compose.yml. that
+#   hasn't worked
+PHP_DEFAULT_TIMEOUT=60
+
+# the variables we will be substituting
+PHP_TIMEOUT=${PHP_TIMEOUT:-${PHP_DEFAULT_TIMEOUT}}
+
+# tell the user what settings we are using
+echo "PHP: ___PHP_TIMEOUT___ is ${PHP_TIMEOUT}"
+
+# the NGINX files we will edit
+NGINX_FILES=$(find /etc/nginx -type f)
+
+# let's edit them
+for x in $NGINX_FILES ; do
+    sed -i "s|___PHP_TIMEOUT___|${PHP_TIMEOUT}|g" $x
+done
+
+# the PHP files we will edit
+PHP_FILES=$(find /etc/php -type f)
+
+# let's edit them
+for x in $PHP_FILES ; do
+    sed -i "s|___PHP_TIMEOUT___|${PHP_TIMEOUT}|g" $x
+done